Supersede `destroy` with `finalizeClassInstance`.
Let us have `finalizeClassInstance` (a function finalizing a class instance referenced by its argument) and `destruct` (a function destroying its argument equivalent as if it goes out of scope). Combining these two functions in one (current `destroy`) is *needless*, *error-prone*, and *inconsistent*.

It is *needless* because having one name assumes is is used in templated code but there is no common uses for such combined function.

It is *error-prone* because e.g. when `destroy` is used in templated code like creating an analog of a struct with manual memory management:
---
void destructIt()
{
    ...
    foreach(i, T; Types)
        destroy(* cast(T*) getMemory(i));
    ...
}
---
it is easy to forget about the fact that `destroy` behaves different for class references. As this is the common situation, `destroy` require a redundant `static if` branch in the bast case and causes nasty bugs in the worst.

It is *inconsistent*, because *druntime* isn't a place for template code at all and really fast and CTFE-able destructing function use templates heavily (see `destruct` proposal for example).