Do not run Fiber unit test on certain systems
Systems which do use the C backend but the C library returns ENOSYS
at runtime. (This happens on ARM with glibc < 2.17 for example).

We do not want to abort all testing in this case.
We also don't want the static ctor to assert in this
case whenever someone wants to use core.thread.