Revert vararg changes in rt.lifetime.
This reverts commits dbfdbb84 and ccc9d3a1, as they were
specific to the "simplified" vararg ABI reverted in the
previous commit. Ideally, the code would require no version
statements at all, and indeed the va_copy should be in the
loop for the program to be correct without making the
assumption that va_list is passed by copy. But as long as
DMD still requires __va_argsave, there is probably little
point in fixing this upstream.