Fix minor memory leak shown in valgrind.
When debuging a trivial program using the memory leak profiler valgrind the d
runtime leaks 88bytes of memory. While this is a small amount that the os will
cleanup anyway on program shutdown, it is a bug nonetheless and is misleading
when looking for real memory leaks in larger programs. The cause of the issue
is that the main thread is not garbage collected so special care has to be
taken to ensure its resources are freed.