core.sys.windows: Allow use on non-Windows platforms
The Windows API binding headers contain declarations which are
occasionally useful even on non-Windows platforms, such as:

- Windows / Portable Executable file format structures, which are used
  for Windows executables, but also portable Common Language
  Infrastructure executables as used by .NET and Mono.

- Bitmap file format structures, used in .bmp image files which can be
  used on any platform.

Limiting the contents of the Windows headers to be available only when
compiling for the Windows platform is generally an artificial
limitation, since one can conceivably think of uses for most
declarations. Implementations of the Win32 API exist on other
platforms and can be used alongside that platform's native APIs (see
e.g. Winelib).

As such, this change removes the `version(Windows):` from the top of
all modules, and instead only limits the `pragma(lib)` linker
directives and Windows-specific helper/wrapper code to the Windows
target platform.