Added code to disable rt_trapExceptions when started inside a debugger on Linux
The implementation is borrowed from Stack Overflow (which is permissively licensed, we must give them attribution which is the same as Boost), and follows the same pattern we do on Windows - if we start in a debugger, don't trap the exception, letting the debugger handle it when it is thrown.

This gives us an interactive debugging session at the throw point, which is of significantly more value than the limited info we get from printing the string.

See also:
http://arsdnet.net/this-week-in-d/2016-aug-07.html