Use OPTIONAL_PIC in the test suite
Test suite didn't compile the executable with fPIC or with -cov no
matter if that was set in the top makefile (since it is a) overriding
DFLAGS, and b) because OPTIONAL_PIC was not exported from the main make
file to sub make files).

This makes test suite compile executables with the fPIC if requested
with PIC. Incidentally, this fixes a problem compiling shared test suite
on arch linux with binutils greater than v2.28 (see issue 17375), since
the executable itself will be compiled with -fPIC, not just shared
library.

N.B.: the -cov flag is present, but not exported and used, because
some shared library tests are segfaulting when compiled with -cov.