Make thread_attachThis nothrow
In case any of the GC operations in the thread_attachThis fail with
the Error, it's likely that GC.enable will also throw an error. This is
not allowed for the statically allocated errors (such as
InvalidMemoryOperationError, commonly thrown by GC), and it will
cause deadlock where the t->next will be same as t.

In order to prevent this, we need to make scope(exit) GC.enable()
scope(success) GC.enable(), but then to make sure there will be no
exception which will cause GC.enable to be skiped, we will make entire
thread_attachThis nothrow.