Fix issue 17914: Disable stack protection page by default
On platforms we're using mmap to allocate the fiber stacks we
could hit the limit of the maximum mappings of the ranges more easily
since the mprotect will create a new range (with the different
protection bits). Hence this sets the default protection page to
non-existent, so we don't regress from the previous behaviour from
v2.074.x and earlier.