Fix llvm_mem* intrinsics for LLVM 7.0.
The signature of llvm.mem* intrinsics have changed: the alignment parameter was removed. This fixes the intrinsics signature, and adds convenience functions for backwards compatibility. Note: the alignment is simply discarded (always valid); so that prevents certain optimizations. Future work could implement setting the `align<n>` attribute on the arguments.