Make errno work with nothrow, @nogc, and @safe code.
For whatever reason, errno wasn't marked with any of those attributes,
and I don't see any reason why it can't be, so this fixes that.

I haven't messed with the functions in the else block, because I don't
understand why it's even there instead of doing what we normally do with
static assert(false, ...) to indicate that the declaration is missing.