Fix Issue 19239 - Fix unsafe casting away of const in hashOf(const Object)
Fixing this will let us legitimately call `hashOf(const Object)` in
`@trusted` code at the expense of a runtime check. We know that the
base `Object.toHash` can be called safely (since it's just based on
the object's address), so we check at runtime that the argument
doesn't override `toHash`. It would be better if we could check at
runtime if `toHash` overridden with a const function, but I don't
believe that is a way to do this.