test: Force Fiber guardPageSize to be same as stackSize
The default initialized hard limit is pulled from sysconf(_SC_PAGESIZE),
however this varies from platform to platform.  On one x86_64-linux
server, the value is 4096, whilst on another powerpc64le-linux server,
it is 65536.  It is for the latter PPC server, that the test failed,
because the Fiber stack is never overflowed.