Fix backward template constraints on `core.atomic.cas`
In the case of `cas`, the first argument is a lvalue, while the other two
can be rvalue. The template constraint was checking whether the lvalue
argument type could be coerced to the rvalue argument type,
however that is backward compared to how things are usually done in the language.
We want rvalue to coerce to rvalue, never the other way around.