Add O_CLOEXEC for glibc linux for x86, arm and ppc
I do not have access to the other arch headers
to check this fully.

The default is 02000000 in glibc / gcc fcntl.h, but each
arch could provide override. The ones added were verified.

I do not want to add other O_* flags at the moment, because
there are many and it is messy to figure out.
For example ppc64e and armhfl and aarch64 have
non-default O_DIRECTORY,
O_NOFOLLOW, O_DIRECT values. And that does not
depend on OS, but is shared for all OSes on that platform.

This was figured out by inspecting content of:

/usr/include/x86_64-linux-gnu/bits/fcntl.h
/usr/include/x86_64-linux-gnu/bits/fcntl-linux.h
/usr/include/i386-linux-gnu/bits/fcntl.h
/usr/include/i386-linux-gnu/bits/fcntl-linux.h
/usr/include/powerpc64le-linux-gnu/bits/fcntl.h
/usr/include/powerpc64le-linux-gnu/bits/fcntl-linux.h
/usr/include/arm-linux-gnueabihf/bits/fcntl.h
/usr/include/arm-linux-gnueabihf/bits/fcntl-linux.h
/usr/include/aarch64-linux-gnu/bits/fcntl.h
/usr/include/aarch64-linux-gnu/bits/fcntl-linux.h