Put attributes on core.time : Duration.toString unittest
The original changes to toString, to support sink delegates,
did not make toString templated, hence the need to check
the GC usage at run time (and the lack of attributes).
However, now that toString is templated, this is no longer needed,
and putting nogc on the unittest is much better.