core.stdcpp.string: char_traits functions shouldn't have implementation
Since `char_traits` is a template, any code within it will have
to be instantiated externaly to be usable.
Our main usage of `char_traits` is as a parameter to other templates,
for which it only matters that its mangling matches.

Moreover, `length` and `move` were breaking `@safe` with prejudice,
taking an arbitrary pointer and doing unsafe things with it.